LPS-55155 Move method to where it is being used to avoid confusion
